Move high and low pass filters to `//filter:filter` lib
Add a `//filter:filter` library with strongly typed high pass and low
pass filters.
