+++ filter/test/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@rules_cc//cc:defs.bzl", "cc_test")
+
+cc_test(
+    name = "filter_test",
+    timeout = "short",
+    srcs = ["filter_test.cpp"],
+    deps = [
+        "//filter",
+        "@boost_ut",
+    ],
+)
